Rework input tests.
