+++ tests/Input/ButtonGroup/RenderTest.php
@@ -14,119 +14,6 @@ use PHPUnit\Framework\TestCase;
  */
 final class RenderTest extends TestCase
 {
-    public function testContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class">
-            <input id="button-6582f2d099e8a" type="submit" value="Submit">
-            <input id="button-6582f2d099e8b" type="reset" value="Reset">
-            </div>
-            HTML,
-            ButtonGroup::widget()
-                ->buttons(
-                    Button::widget()->id('button-6582f2d099e8a')->type('submit')->value('Submit'),
-                    Button::widget()->id('button-6582f2d099e8b')->type('reset')->value('Reset')
-                )
-                ->containerAttributes(['class' => 'class'])
-                ->render()
-        );
-    }
-
-    public function testContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class">
-            <input id="button-6582f2d099e8a" type="submit" value="Submit">
-            <input id="button-6582f2d099e8b" type="reset" value="Reset">
-            </div>
-            HTML,
-            ButtonGroup::widget()
-                ->buttons(
-                    Button::widget()->id('button-6582f2d099e8a')->type('submit')->value('Submit'),
-                    Button::widget()->id('button-6582f2d099e8b')->type('reset')->value('Reset')
-                )
-                ->containerClass('class')
-                ->render()
-        );
-    }
-
-    public function testContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <article>
-            <input id="button-6582f2d099e8a" type="submit" value="Submit">
-            <input id="button-6582f2d099e8b" type="reset" value="Reset">
-            </article>
-            HTML,
-            ButtonGroup::widget()
-                ->buttons(
-                    Button::widget()->id('button-6582f2d099e8a')->type('submit')->value('Submit'),
-                    Button::widget()->id('button-6582f2d099e8b')->type('reset')->value('Reset')
-                )
-                ->containerTag('article')
-                ->render()
-        );
-    }
-
-    public function testContainerWithFalse(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="button-6582f2d099e8a" type="submit" value="Submit">
-            <input id="button-6582f2d099e8b" type="reset" value="Reset">
-            HTML,
-            ButtonGroup::widget()
-                ->buttons(
-                    Button::widget()->id('button-6582f2d099e8a')->type('submit')->value('Submit'),
-                    Button::widget()->id('button-6582f2d099e8b')->type('reset')->value('Reset')
-                )
-                ->container(false)
-                ->render()
-        );
-    }
-
-    public function testIndividualContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div>
-            <div>
-            <input id="button-6582f2d099e8a" type="submit" value="Submit">
-            </div>
-            <div>
-            <input id="button-6582f2d099e8b" type="reset" value="Reset">
-            </div>
-            </div>
-            HTML,
-            ButtonGroup::widget()
-                ->buttons(
-                    Button::widget()->id('button-6582f2d099e8a')->type('submit')->value('Submit'),
-                    Button::widget()->id('button-6582f2d099e8b')->type('reset')->value('Reset')
-                )
-                ->individualContainer(true)
-                ->render()
-        );
-    }
-
-    public function testLoadDefaultDefinitionWithContainerWithFalse(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="button-6582f2d099e8a" type="submit" value="Submit">
-            <input id="button-6582f2d099e8b" type="reset" value="Reset">
-            HTML,
-            ButtonGroup::widget(['container()' => [false]])
-                ->buttons(
-                    Button::widget()->id('button-6582f2d099e8a')->type('submit')->value('Submit'),
-                    Button::widget()->id('button-6582f2d099e8b')->type('reset')->value('Reset')
-                )
-                ->render()
-        );
-    }
-
     public function testRender(): void
     {
         Assert::equalsWithoutLE(
