+++ tests/Input/CheckboxList/AttributesTest.php
@@ -12,23 +12,23 @@ use PHPUnit\Framework\TestCase;
 /**
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class RenderTest extends TestCase
+final class AttributesTest extends TestCase
 {
     public function testAriaDescribedBy(): void
     {
         Assert::equalsWithoutLE(
             <<<HTML
             <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1" aria-describedby="MyWidget">
+            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1" aria-describedby="value">
             <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2" aria-describedby="MyWidget">
+            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2" aria-describedby="value">
             <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3" aria-describedby="MyWidget">
+            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3" aria-describedby="value">
             <label for="checkbox-6599b6a33dd97">Orange</label>
             </div>
             HTML,
             CheckboxList::widget()
-                ->ariaDescribedBy('MyWidget')
+                ->ariaDescribedBy('value')
                 ->id('checkboxlist-65858c272ea89')
                 ->items(
                     Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
@@ -36,7 +36,7 @@ final class RenderTest extends TestCase
                     Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
                 )
                 ->name('CheckboxForm[text]')
-                ->render(),
+                ->render()
         );
     }
 
@@ -45,16 +45,16 @@ final class RenderTest extends TestCase
         Assert::equalsWithoutLE(
             <<<HTML
             <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1" aria-label="MyWidget">
+            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1" aria-label="value">
             <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2" aria-label="MyWidget">
+            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2" aria-label="value">
             <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3" aria-label="MyWidget">
+            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3" aria-label="value">
             <label for="checkbox-6599b6a33dd97">Orange</label>
             </div>
             HTML,
             CheckboxList::widget()
-                ->ariaLabel('MyWidget')
+                ->ariaLabel('value')
                 ->id('checkboxlist-65858c272ea89')
                 ->items(
                     Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
@@ -62,7 +62,7 @@ final class RenderTest extends TestCase
                     Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
                 )
                 ->name('CheckboxForm[text]')
-                ->render(),
+                ->render()
         );
     }
 
@@ -71,16 +71,16 @@ final class RenderTest extends TestCase
         Assert::equalsWithoutLE(
             <<<HTML
             <div id="checkboxlist-65858c272ea89">
-            <input class="class" id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
+            <input class="value" id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
             <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input class="class" id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
+            <input class="value" id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
             <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input class="class" id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
+            <input class="value" id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
             <label for="checkbox-6599b6a33dd97">Orange</label>
             </div>
             HTML,
             CheckboxList::widget()
-                ->attributes(['class' => 'class'])
+                ->attributes(['class' => 'value'])
                 ->id('checkboxlist-65858c272ea89')
                 ->items(
                     Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
@@ -118,145 +118,73 @@ final class RenderTest extends TestCase
         );
     }
 
+    public function testCheckedValue(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <div id="checkboxlist-65858c272ea89">
+            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1" checked>
+            <label for="checkbox-6599b6a33dd96">Apple</label>
+            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2" checked>
+            <label for="checkbox-6599b6a33dd98">Banana</label>
+            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
+            <label for="checkbox-6599b6a33dd97">Orange</label>
+            </div>
+            HTML,
+            CheckboxList::widget()
+                ->checkedValue([1, 2])
+                ->id('checkboxlist-65858c272ea89')
+                ->items(
+                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
+                )
+                ->name('CheckboxForm[text]')
+                ->render(),
+        );
+    }
+
+    public function testCheckedValueWithNull(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <div id="checkboxlist-65858c272ea89">
+            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
+            <label for="checkbox-6599b6a33dd96">Apple</label>
+            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
+            <label for="checkbox-6599b6a33dd98">Banana</label>
+            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
+            <label for="checkbox-6599b6a33dd97">Orange</label>
+            </div>
+            HTML,
+            CheckboxList::widget()
+                ->checkedValue(null)
+                ->id('checkboxlist-65858c272ea89')
+                ->items(
+                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
+                )
+                ->name('CheckboxForm[text]')
+                ->render(),
+        );
+    }
+
     public function testClass(): void
     {
         Assert::equalsWithoutLE(
             <<<HTML
             <div id="checkboxlist-65858c272ea89">
-            <input class="class" id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
+            <input class="value" id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
             <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input class="class" id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
+            <input class="value" id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
             <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input class="class" id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
+            <input class="value" id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
             <label for="checkbox-6599b6a33dd97">Orange</label>
             </div>
             HTML,
             CheckboxList::widget()
-                ->class('class')
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class" id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->containerAttributes(['class' => 'class'])
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class" id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->containerClass('class')
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <article id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </article>
-            HTML,
-            CheckboxList::widget()
-                ->containerTag('article')
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testContainerWithFalse(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            HTML,
-            CheckboxList::widget()
-                ->container(false)
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testEnclosedByLabel(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div id="checkboxlist-65858c272ea89">
-            <label for="checkbox-6599b6a33dd96"><input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">Apple</label>
-            <label for="checkbox-6599b6a33dd98"><input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">Banana</label>
-            <label for="checkbox-6599b6a33dd97"><input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->enclosedByLabel(true)
+                ->class('value')
                 ->id('checkboxlist-65858c272ea89')
                 ->items(
                     Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
@@ -294,11 +222,38 @@ final class RenderTest extends TestCase
         );
     }
 
+
+    public function testGenerateAriaDescribedByWithFalse(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <div id="checkboxlist-65858c272ea89">
+            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
+            <label for="checkbox-6599b6a33dd96">Apple</label>
+            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
+            <label for="checkbox-6599b6a33dd98">Banana</label>
+            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
+            <label for="checkbox-6599b6a33dd97">Orange</label>
+            </div>
+            HTML,
+            CheckboxList::widget()
+                ->ariaDescribedBy(false)
+                ->id('checkboxlist-65858c272ea89')
+                ->items(
+                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
+                )
+                ->name('CheckboxForm[text]')
+                ->render(),
+        );
+    }
+
     public function testId(): void
     {
         Assert::equalsWithoutLE(
             <<<HTML
-            <div id="id">
+            <div id="value">
             <input id="id-checkbox-1" name="CheckboxForm[text][]" type="checkbox" value="1">
             <label for="id-checkbox-1">Apple</label>
             <input id="id-checkbox-2" name="CheckboxForm[text][]" type="checkbox" value="2">
@@ -308,7 +263,7 @@ final class RenderTest extends TestCase
             </div>
             HTML,
             CheckboxList::widget()
-                ->id('id')
+                ->id('value')
                 ->items(
                     Checkbox::widget()->id('id-checkbox-1')->labelContent('Apple')->value(1),
                     Checkbox::widget()->id('id-checkbox-2')->labelContent('Banana')->value(2),
@@ -319,128 +274,16 @@ final class RenderTest extends TestCase
         );
     }
 
-    public function testLabel(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <label for="checkboxlist-65858c272ea89">Select your fruits?</label>
-            <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->labelContent('Select your fruits?')
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testLabelAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <label class="class" for="checkboxlist-65858c272ea89">Select your fruits?</label>
-            <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->labelAttributes(['class' => 'class'])
-                ->labelContent('Select your fruits?')
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testLabelClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <label class="class" for="checkboxlist-65858c272ea89">Select your fruits?</label>
-            <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->labelClass('class')
-                ->labelContent('Select your fruits?')
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
-    public function testLabelItemClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <label for="checkboxlist-65858c272ea89">Select your fruits?</label>
-            <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label class="class" for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label class="class" for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label class="class" for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->labelContent('Select your fruits?')
-                ->labelItemClass('class')
-                ->name('CheckboxForm[text]')
-                ->render(),
-        );
-    }
-
     public function testName(): void
     {
         Assert::equalsWithoutLE(
             <<<HTML
-            <label class="class" for="checkboxlist-65858c272ea89">Select your fruits?</label>
             <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
+            <input id="checkbox-6599b6a33dd96" name="value[]" type="checkbox" value="1">
             <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
+            <input id="checkbox-6599b6a33dd98" name="value[]" type="checkbox" value="2">
             <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
+            <input id="checkbox-6599b6a33dd97" name="value[]" type="checkbox" value="3">
             <label for="checkbox-6599b6a33dd97">Orange</label>
             </div>
             HTML,
@@ -451,9 +294,7 @@ final class RenderTest extends TestCase
                     Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
                     Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
                 )
-                ->labelClass('class')
-                ->labelContent('Select your fruits?')
-                ->name('CheckboxForm[text]')
+                ->name('value')
                 ->render(),
         );
     }
@@ -478,7 +319,6 @@ final class RenderTest extends TestCase
                     Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
                     Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
                 )
-                ->labelClass('class')
                 ->labelContent('Select your gender?')
                 ->name('CheckboxForm[text]')
                 ->notLabel()
@@ -559,63 +399,6 @@ final class RenderTest extends TestCase
                 )
                 ->name('CheckboxForm[text]')
                 ->tabIndex(1)
-                ->render(),
-        );
-    }
-
-    public function testTemplate(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div id="checkboxlist-65858c272ea89">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            <label class="class" for="checkboxlist-65858c272ea89">Select your fruits?</label>
-            HTML,
-            CheckboxList::widget()
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->labelClass('class')
-                ->labelContent('Select your fruits?')
-                ->name('CheckboxForm[text]')
-                ->template('{tag}\n{label}')
-                ->render(),
-        );
-    }
-
-    public function testUncheckValue(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div id="checkboxlist-65858c272ea89" tabindex="1">
-            <input name="CheckboxForm[text][]" type="hidden" value="0">
-            <input id="checkbox-6599b6a33dd96" name="CheckboxForm[text][]" type="checkbox" value="1">
-            <label for="checkbox-6599b6a33dd96">Apple</label>
-            <input id="checkbox-6599b6a33dd98" name="CheckboxForm[text][]" type="checkbox" value="2">
-            <label for="checkbox-6599b6a33dd98">Banana</label>
-            <input id="checkbox-6599b6a33dd97" name="CheckboxForm[text][]" type="checkbox" value="3">
-            <label for="checkbox-6599b6a33dd97">Orange</label>
-            </div>
-            HTML,
-            CheckboxList::widget()
-                ->id('checkboxlist-65858c272ea89')
-                ->items(
-                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
-                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
-                )
-                ->name('CheckboxForm[text]')
-                ->tabIndex(1)
-                ->uncheckValue('0')
                 ->render(),
         );
     }
@@ -741,6 +524,55 @@ final class RenderTest extends TestCase
                     Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
                 )
                 ->name('CheckboxForm[text]')
+                ->render(),
+        );
+    }
+
+    public function testWithoutId(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <div>
+            <input id="checkbox-6599b6a33dd96" type="checkbox" value="1">
+            <label for="checkbox-6599b6a33dd96">Apple</label>
+            <input id="checkbox-6599b6a33dd98" type="checkbox" value="2">
+            <label for="checkbox-6599b6a33dd98">Banana</label>
+            <input id="checkbox-6599b6a33dd97" type="checkbox" value="3">
+            <label for="checkbox-6599b6a33dd97">Orange</label>
+            </div>
+            HTML,
+            CheckboxList::widget()
+                ->id(null)
+                ->items(
+                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
+                )
+                ->render(),
+        );
+    }
+
+    public function testWithoutName(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <div id="checkboxlist-65858c272ea89">
+            <input id="checkbox-6599b6a33dd96" type="checkbox" value="1">
+            <label for="checkbox-6599b6a33dd96">Apple</label>
+            <input id="checkbox-6599b6a33dd98" type="checkbox" value="2">
+            <label for="checkbox-6599b6a33dd98">Banana</label>
+            <input id="checkbox-6599b6a33dd97" type="checkbox" value="3">
+            <label for="checkbox-6599b6a33dd97">Orange</label>
+            </div>
+            HTML,
+            CheckboxList::widget()
+                ->id('checkboxlist-65858c272ea89')
+                ->items(
+                    Checkbox::widget()->id('checkbox-6599b6a33dd96')->labelContent('Apple')->value(1),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd98')->labelContent('Banana')->value(2),
+                    Checkbox::widget()->id('checkbox-6599b6a33dd97')->labelContent('Orange')->value(3),
+                )
+                ->name(null)
                 ->render(),
         );
     }
