--- tests/Methods}.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PHPForge\Html\Tests;
-
-use PHPForge\Html\Input\Image;
-use PHPUnit\Framework\TestCase;
-
-final class MethodsTest extends TestCase
-{
-    public function testPrintMethods(): void
-    {
-        $methods = get_class_methods(Image::class);
-        sort($methods);
-        var_dump(implode("\n", $methods));
-        die;
-    }
-}
